# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CastleRock Investments — Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /import/preview:
    post:
      operationId: previewImport
      tags: [import]
      summary: Parse CSV and return preview rows (no save)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportUpload"
      responses:
        "200":
          description: Preview result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreview"
  /import/confirm:
    post:
      operationId: confirmImport
      tags: [import]
      summary: Save parsed trades from a preview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportConfirm"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
components:
  schemas:
...
    ImportUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
        brokerFormat:
          type: string
          description: "e.g. mt4, mt5, generic, ibkr"
        accountId:
          type: ["integer","null"]
      required: [file]
    ImportPreview:
      type: object
      properties:
        sessionId: { type: string }
        detectedFormat: { type: string }
        columnMapping:
          type: object
          additionalProperties: { type: string }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
        totalRows: { type: integer }
        validRows: { type: integer }
        duplicateRows: { type: integer }
      required: [sessionId, detectedFormat, columnMapping, rows, errors, totalRows, validRows, duplicateRows]
    ImportRow:
      type: object
      properties:
        rowIndex: { type: integer }
        symbol: { type: string }
        direction: { type: string }
        entryDate: { type: string }
        exitDate: { type: ["string","null"] }
        entryPrice: { type: number }
        exitPrice: { type: ["number","null"] }
        quantity: { type: number }
        pnl: { type: ["number","null"] }
        isDuplicate: { type: boolean }
        raw:
          type: object
          additionalProperties: { type: string }
      required: [rowIndex, symbol, direction, entryDate, entryPrice, quantity, isDuplicate, raw]
    ImportError:
      type: object
      properties:
        rowIndex: { type: integer }
        field: { type: string }
        message: { type: string }
      required: [rowIndex, field, message]
    ImportConfirm:
      type: object
      properties:
        sessionId: { type: string }
        accountId: { type: ["integer","null"] }
        skipDuplicates: { type: boolean }
        brokerFormat: { type: string }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ImportRow"
      required: [sessionId, rows]
    ImportResult:
      type: object
      properties:
        batchId: { type: integer }
        savedCount: { type: integer }
        skippedCount: { type: integer }
        errorCount: { type: integer }
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
      required: [batchId, savedCount, skippedCount, errorCount, errors]
...