# lib/api-zod/src/generated/types/importError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CastleRock Investments — Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportError {
  rowIndex: number;
  field: string;
  message: string;
}